Testing background process with Dash Pages
